updates to parallel downlad (minus sunpy) 1. add a script for testing rather than a notebook 2. change from processes to threads for downloads (don't need a lot of CPU power) 3. remove logging because it's annoying 4. add attributes to the DownloadResults 5. change a little bit of logic

+++ net/aia_requests.py
@@ -1,11 +1,13 @@
 from . import aia_fmt_xml as afx
+from dataclasses import dataclass
 
 import astropy.units as u
 import astropy.time
+import copy
 import functools
-import logging
 import math
 import multiprocessing as mp
+import multiprocessing.dummy
 import os
 import requests
 import requests.exceptions
@@ -13,13 +15,22 @@ import sys
 import typing
 import xmltodict
 
-# configure this to be as you want
-read_timeout = 10 << u.s
+@dataclass
+class Config:
+    read_timeout: u.s
+    debug: bool
 
+# configure this to be as you want
+cfg = Config(10 << u.s, debug=False)
+def debug_print(*args, **kwargs):
+    if cfg.debug:
+        print(*args, **kwargs)
 
 class DownloadResult(typing.NamedTuple):
+    url: str
     file: str
     success: bool
+    error: Exception
 
 
 @u.quantity_input
@@ -38,34 +49,38 @@ def download_aia_between(
     files_downloaded = []
     all_urls = []
     # do this sequentially because it's fast
-    logging.debug('start find aia urls')
+    debug_print('start find aia urls')
     for w in wavelengths:
         all_urls += build_aia_urls(start, end, w)
-    logging.debug(f'done find aia urls:\n{all_urls}')
+    debug_print(f'done find aia urls:\n{all_urls}')
 
     download_wrapper = functools.partial(actual_download_files, fits_out_dir)
 
     # do this in parallel because it's slow
     tries = 0
     while tries < attempts:
-        logging.debug('start try downloads')
-        with mp.Pool(processes=num_jobs) as p:
+        debug_print('start try downloads')
+        with mp.dummy.Pool(processes=num_jobs) as p:
             files_downloaded += p.map(
                 download_wrapper,
                 all_urls,
                 chunksize=math.ceil(len(all_urls) / num_jobs)
             )
-        logging.debug('done try downloads')
+        debug_print('done try downloads')
 
-        errored = [all_urls[i] for (i, res) in enumerate(files_downloaded) if not res.success]
-        if not errored: break
+        all_downloads_worked = all(res.success for res in files_downloaded)
+        if all_downloads_worked: break
+        else:
+            if cfg.debug:
+                debug_print('some failed ones:\n', [f for f in files_downloaded if not f.success])
 
+        errored = [res.url for res in files_downloaded if not res.success]
         failed = len(errored)
         initial = len(all_urls)
         errored_to_print = '\n' + '\t\n'.join(errored)
-        logging.info(f'{failed} / {initial} downloads failed.')
-        logging.info(f'retrying the following (attempt {1 + tries} / {attempts}): {errored_to_print}')
-        all_urls = errored
+        debug_print(f'{failed} / {initial} downloads failed.')
+        debug_print(f'retrying the following (attempt {1 + tries} / {attempts}): {errored_to_print}')
+        all_urls = copy.deepcopy(errored)
         tries += 1
 
     return files_downloaded
@@ -178,27 +193,37 @@ def actual_download_files(output_directory: str, url: str) -> DownloadResult:
     download AIA files
     return: (file name output, success or not)
     '''
-    logging.debug('gotcha:', url)
+    debug_print(f'wait time is {cfg.read_timeout}')
+    debug_print('gotcha:', url)
     full_fn = '🥲'
     try:
-        with requests.get(url, stream=True, timeout=read_timeout.to(u.s).value) as res:
+        with requests.get(url, stream=True, timeout=cfg.read_timeout.to(u.s).value) as res:
             res.raise_for_status()
             fn = parse_filename(res.headers['Content-Disposition'])
+            file_size = int(res.headers['Content-Length'])
+
             full_fn = f'{output_directory}/{fn}'
+            if os.path.exists(full_fn) and os.stat(full_fn):
+                print('file already exists and is downloaded:', full_fn)
+                debug_print()
+                return DownloadResult(url, full_fn, True, None)
+
             with open(full_fn, 'wb') as f:
                 for chunk in res.iter_content(chunk_size=1024 * 1024):
                     f.write(chunk)
 
-    except requests.exceptions.ReadTimeout:
-        return DownloadResult(full_fn, False)
+    except requests.exceptions.ReadTimeout as e:
+        return DownloadResult(url, full_fn, False, e)
+    except requests.exceptions.HTTPError as e:
+        return DownloadResult(url, full_fn, False, e)
 
-    return DownloadResult(full_fn, True)
+    return DownloadResult(url, full_fn, True, None)
 
 
 def test():
     print('request test')
-    start = astropy.time.Time('2020-02-01T03:00:00')
-    end = astropy.time.Time('2020-02-01T03:01:00')
+    start = astropy.time.Time('2019-04-03T17:40:00.0')
+    end = astropy.time.Time('2019-04-03T17:55:00.0')
     wavelengths = [171 * u.Angstrom]
 
     out_dir = 'test-manual-download'
@@ -209,14 +234,17 @@ def test():
         wavelengths=wavelengths,
         fits_out_dir=out_dir,
         num_jobs=8,
-        attempts=5
+        attempts=10
     )
+    if all(r.success for r in ret):
+        print('all good')
+    else:
+        print('some failed')
     return ret
 
 
 def timed_test():
     from datetime import datetime
-    logging.basicConfig(stream=sys.stdout, level=logging.DEBUG)
 
     start_time = datetime.now()
     ret = test()
